backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bs4==0.0.2
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.129.0
geopy==2.4.1
h11==0.16.0
httpx==0.28.1
idna==3.11
numpy==2.2.6
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==22.0.0
pydantic==2.12.5
email-validator==2.3.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.46
starlette==0.52.1
supabase==2.28.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
firebase-admin
bcrypt
python-jose
